fix review service

+++ src/main/java/com/nam/storespring/service/ReviewService.java
@@ -2,7 +2,6 @@ package com.nam.storespring.service;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.nam.storespring.dto.request.ReviewRequest;
@@ -11,13 +10,23 @@ import com.nam.storespring.entity.Review;
 import com.nam.storespring.mapper.ReviewMapper;
 import com.nam.storespring.repository.ReviewRepository;
 
+import lombok.extern.slf4j.Slf4j;
+
 @Service
+@Slf4j
 public class ReviewService {
-    @Autowired
     private ReviewRepository reviewRepository;
 
-    @Autowired
     private ReviewMapper reviewMapper;
+
+    private AuthenticationService authenticationService;
+
+    public ReviewService(AuthenticationService authenticationService, ReviewRepository reviewRepository,
+            ReviewMapper reviewMapper) {
+        this.authenticationService = authenticationService;
+        this.reviewMapper = reviewMapper;
+        this.reviewRepository = reviewRepository;
+    }
 
     public List<Review> getReviews() {
         return reviewRepository.findAll();
@@ -26,10 +35,12 @@ public class ReviewService {
     public ReviewResponse createReview(ReviewRequest request) {
         Review review = new Review();
 
-        review.setUserId(request.getUserId());
+        review.setUserId(authenticationService.getCurrentUserId());
         review.setProductId(request.getProductId());
         review.setRating(request.getRating());
         review.setComment(request.getComment());
+
+        log.info("User ID from token: {}", authenticationService.getCurrentUserId());
 
         return reviewMapper.toReviewResponse(reviewRepository.save(review));
     }
@@ -38,6 +49,12 @@ public class ReviewService {
         Review review = reviewRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Review not found"));
 
+        String currentUserId = authenticationService.getCurrentUserId();
+
+        if (!review.getUserId().equals(currentUserId)) {
+            throw new RuntimeException("You do not have permission to update this review");
+        }
+
         review.setComment(request.getComment());
 
         return reviewMapper.toReviewResponse(reviewRepository.save(review));
@@ -45,6 +62,15 @@ public class ReviewService {
     }
 
     public void deleteReview(String reviewId) {
+        Review review = reviewRepository.findById(reviewId)
+                .orElseThrow(() -> new IllegalArgumentException("Review not found"));
+
+        String currentUserId = authenticationService.getCurrentUserId();
+
+        if (!review.getUserId().equals(currentUserId)
+                && !currentUserId.equals("813d9125-b8c7-46ac-8eb6-148a2bcf02c4")) {
+            throw new RuntimeException("You do not have permission to delete this review");
+        }
         reviewRepository.deleteById(reviewId);
     }
 }
